GK_LU_refinements
What you have is *almost* perfect.  This code proposes a couple of minor changes.  One is at the start where you want users to enter an integer yet you almost invite them to enter a float by saying between $1.00 and $10.00 instead of between $1 and $10.

Note that when you use the 'return' command in a function, that stops the function which means resetting your 'valid' variable to True is not necessary in that case.